Trekk ut slåSammenFinnmarkstillegg til egen fil
