+++ pom.xml
@@ -24,7 +24,7 @@
         <sha1/>
         <changelist>-SNAPSHOT</changelist>
         <prosessering.version>2.20250825143618_c0302c9</prosessering.version>
-        <felles.version>3.20250804101327_a0f5fad</felles.version>
+        <felles.version>3.20250915092837_153983f</felles.version>
         <eksterne-kontrakter-bisys.version>2.0_20250826113118_c8aa66d</eksterne-kontrakter-bisys.version>
         <felles-kontrakter.version>3.0_20250902143555_da31aa4</felles-kontrakter.version>
         <familie.kontrakter.saksstatistikk>2.0_20250826113118_c8aa66d</familie.kontrakter.saksstatistikk>
